app/requirements.txt
playwright
playwright-stealth
boto3
python-whois
pymysql
python-dotenv
cryptography